backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.7.4
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
langchain>=0.2.16
langsmith>=0.1.147
google-cloud-aiplatform>=1.49.0
google-cloud-storage>=2.16.0
google-generativeai>=0.7.0
vertexai>=1.49.0
pypdf>=4.2.0
pdfplumber>=0.11.0
pymupdf>=1.24.3
httpx>=0.27.0
Pillow>=10.3.0
